backend/requirements.txt
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.2
dnspython==2.3.0
Flask==2.2.3
Flask-Cors==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycparser==2.21
PyJWT==2.6.0
pymongo==4.3.3
python-dotenv==1.0.0
Werkzeug==2.2.3
